Time: 0 ms (100%), Space: 8.5 MB (53.19%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
         if(x<0){
             return false;
         }
-        size_t reverse=0;
+        long reverse=0;
         int original =x;
         while (x!=0){
             int num = x %10; // last digit of the number!
